fix: ensure vec_ids are cast to int64 when creating candidate items in search function

+++ IVF.py
@@ -191,10 +191,10 @@ def search(vec_db, query_vector, k=5, batch_size=500):
 
         vecs = vecs / (np.linalg.norm(vecs, axis=1, keepdims=True) + 1e-12)
 
-        for vid, vec in zip(vec_ids.astype(np.int64), vecs):
+        for vid, vec in zip(vec_ids, vecs):
             score = cal_score(query_vector, vec)
 
-            item = (score, -vid)
+            item = (score, -(vid.astype(np.int64)))
 
             if len(candidates) < k:
                 heapq.heappush(candidates, item)
